Move block functions to distinct packages
