+++ go.mod
@@ -1,0 +1,10 @@
+module gobar
+
+go 1.17
+
+require go.i3wm.org/i3 v0.0.0-20190720062127-36e6ec85cc5a
+
+require (
+	github.com/BurntSushi/xgb v0.0.0-20160522181843-27f122750802 // indirect
+	github.com/BurntSushi/xgbutil v0.0.0-20160919175755-f7c97cef3b4e // indirect
+)
