+++ internal/blockutils/colors.go
@@ -1,0 +1,16 @@
+package blockutils
+
+// Color - Create a type to store our color strings.
+type Color string
+
+// Create an enum of colors
+const (
+	Black   Color = `#222222`
+	Red     Color = `#e84f4f`
+	Green   Color = `#b7ce42`
+	Blue    Color = `#66aabb`
+	Magenta Color = `#b7416e`
+	Cyan    Color = `#6d878d`
+	White   Color = `#dddddd`
+	None    Color = ``
+)
