+++ internal/media/media.go
@@ -1,16 +1,11 @@
-package main
+package media
 
 import (
+	"gobar/internal/blockutils"
+	"gobar/internal/log"
 	"os/exec"
 	"time"
 )
-
-type mediaState struct {
-	lastPlayer string
-	text       string
-	head       int
-	tail       int
-}
 
 //TODO: This is lazy - I should find a better way to do this eventually.
 var state = mediaState{
@@ -20,39 +15,11 @@ var state = mediaState{
 	tail:       -1,
 }
 
-func (state *mediaState) scroll() string {
-	if len(state.text) > 50 {
-		if state.head == state.tail {
-			state.tail = 46
-		}
-
-		if state.tail+1 == len(state.text) {
-			state.tail = -1
-		}
-
-		if state.head+1 == len(state.text) {
-			state.head = -1
-		}
-
-		state.tail++
-		state.head++
-
-		if state.tail < state.head {
-			return state.text[state.head:] + "   " + state.text[:state.tail]
-		}
-
-		return state.text[state.head:state.tail] + "..."
-	}
-	state.head = -1
-	state.tail = -1
-	return state.text
-}
-
 func getPlayers() []string {
 	cmd := exec.Command("playerctl", "-l")
-	out, err := runCmdStdout(cmd)
+	out, err := blockutils.RunCmdStdout(cmd)
 	if err != nil {
-		fileLog(err)
+		log.FileLog(err)
 	}
 	return out
 }
@@ -60,9 +27,9 @@ func getPlayers() []string {
 func getCurPlayer(players []string) string {
 	for _, player := range players {
 		stateCmd := exec.Command("playerctl", "-p", player, "status")
-		state, err := runCmdStdout(stateCmd)
+		state, err := blockutils.RunCmdStdout(stateCmd)
 		if err != nil {
-			fileLog(err)
+			log.FileLog(err)
 		}
 		if state[0] == "Playing" {
 			return player
@@ -72,9 +39,9 @@ func getCurPlayer(players []string) string {
 }
 
 func getPlayerState(player string) string {
-	status, err := runCmdStdout(exec.Command("playerctl", "-p", player, "status"))
+	status, err := blockutils.RunCmdStdout(exec.Command("playerctl", "-p", player, "status"))
 	if err != nil {
-		fileLog(err)
+		log.FileLog(err)
 	}
 	if len(status) > 0 {
 		if status[0] == "Playing" {
@@ -85,11 +52,12 @@ func getPlayerState(player string) string {
 	return string('\uf04b')
 }
 
-func getMedia(timeout time.Duration, blockCh chan<- *block) {
+// GetMedia sends media information for the media block
+func GetMedia(timeout time.Duration, blockCh chan<- *blockutils.Block) {
 	fmtStr := `{{ artist }} - {{ title }}`
-	mediaBlock := block{
-		Name:        MediaName,
-		Border:      Red,
+	mediaBlock := blockutils.Block{
+		Name:        blockutils.MediaName,
+		Border:      blockutils.Red,
 		BorderLeft:  0,
 		BorderRight: 0,
 		BorderTop:   0,
@@ -106,9 +74,9 @@ func getMedia(timeout time.Duration, blockCh chan<- *block) {
 
 		if state.lastPlayer != "" {
 			infoCmd := exec.Command("playerctl", "-p", state.lastPlayer, "metadata", "-f", fmtStr)
-			curState, err := runCmdStdout(infoCmd)
+			curState, err := blockutils.RunCmdStdout(infoCmd)
 			if err != nil {
-				fileLog(err)
+				log.FileLog(err)
 				continue
 			}
 
@@ -129,21 +97,23 @@ func getMedia(timeout time.Duration, blockCh chan<- *block) {
 	}
 }
 
-func clickMedia(evt *click) {
+/*
+// ClickMedia handles click events for the media block.
+func ClickMedia(evt *clickutils.Click) {
 	action := ""
 
 	switch evt.Button {
-	case leftClick:
+	case clickutils.LeftClick:
 		action = "play-pause"
-	case scrollUp:
+	case clickutils.ScrollUp:
 		action = "previous"
-	case scrollDown:
+	case clickutils.ScrollDown:
 		action = "next"
 	}
 
 	cmd := exec.Command("playerctl", "-p", state.lastPlayer, action)
 
 	if err := cmd.Run(); err != nil {
-		fileLog("Could not control media:", err)
+		log.FileLog("Could not control media:", err)
 	}
-}
+}*/
